[admin] move log reads into backends
